fix(appliance): fix service selector
It was already using simple `app: sourcegraph-appliance`, but this label
wasn't on the pod template. Add it, and also remove the rest of the
service selector just because we can. This simplifies the code we need
to write on the appliance side.

Also simplify the deployment metadata.name, so that it can be
well-known, independent of the helm release name, like the other
resources here (since the release will be a singleton in namespace scope).