frontend: specify databaseAuth consistently first in env
In all places we include the databaseAuth we add it first in the env
list, except for the frontend pod. This updates it to be consistently
first.

Context: For multitenant deployments we want to override the DB user the
frontend uses to connect such that it is different to the migrator. This
allows us to do that.