feat: add batch-change release.yaml for automation (#703)
Since we're dependent on batch-changes for pr creation now (given it
doesn't clone the .git files so we can't create PRs from execution
phase) we have to modify the release.yaml. Instead of breaking existing
workflows, we're creating a new release.yaml here eventually to be
deprecated. I tested this yesterday on the docker repo, and it worked
fine, so I'm replicating that here.

It's a direct copy of the `release.yaml`, except since it's in a subdir,
it needs to change the directory `sg ops update-images` runs from. I
also had to remove references to `git` and `gh`
### Checklist

- [ ] Follow the [manual testing
process](https://github.com/sourcegraph/deploy-sourcegraph-helm/blob/main/TEST.md)
- [ ] Update
[changelog](https://github.com/sourcegraph/deploy-sourcegraph-helm/blob/main/charts/sourcegraph/CHANGELOG.md)
- [ ] Update [Kubernetes update
doc](https://docs.sourcegraph.com/admin/updates/kubernetes)

### Test plan
Run batch change temporal workflow

<!--
As part of SOC2/GN-104 and SOC2/GN-105 requirements, all pull requests
are REQUIRED to
provide a "test plan". A test plan is a loose explanation of what you
have done or
implemented to test this, as outlined in our Testing principles and
guidelines:

https://docs.sourcegraph.com/dev/background-information/testing_principles
  Write your test plan here after the "Test plan" header.
-->